pyproject.toml
[project]
name = "mcp-streamddeck"
version = "0.1.0"
description = "Streamlit MCP Config & Chat Interface"
authors = [
    {name = "Bongki Lee",email = "qhdrl12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "streamlit>=1.44.0",
    "langchain>=0.3.21",
    "langchain-community>=0.3.20",
    "openai>=1.69.0",
    "anthropic>=0.49.0",
    "python-dotenv>=1.1.0",
    "jsonschema>=4.23.0",
    "langchain-mcp-adapters>=0.0.5",
    "mcp>=0.0.1",
    "langchain-google-genai (>=2.1.2,<3.0.0)",
    "langchain-anthropic (>=0.3.10,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "langgraph (>=0.3.21,<0.4.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "langgraph-prebuilt (>=0.1.8,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

